+++ src/python/zensols/util/__init__.py
@@ -1,3 +1,4 @@
+from .writable import *
 from .fail import *
 from .std import *
 from .time import *