+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [1.15.1] - 2025-01-23
+### Changed
+- Move `zensols.config.Writable` to `zensols.util`.  The class is still
+  imported to the `zensols.config` module so import statements need not change.
+- `zensols.util.Failure` extends from `zensols.util.Writable`.
+
+
 ## [1.15.0] - 2025-01-11
 Additional feature additions to configuration. One of these feature additions
 precludes support of Python 3.10.  All dependent packages can continue to use
@@ -628,7 +635,8 @@ Major release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/util/compare/v1.15.0...HEAD
+[Unreleased]: https://github.com/plandes/util/compare/v1.15.1...HEAD
+[1.15.1]: https://github.com/plandes/util/compare/v1.15.0...v1.15.1
 [1.15.0]: https://github.com/plandes/util/compare/v1.14.5...v1.15.0
 [1.14.5]: https://github.com/plandes/util/compare/v1.14.4...v1.14.5
 [1.14.4]: https://github.com/plandes/util/compare/v1.14.3...v1.14.4